requirements.txt
asciiartist==0.5.3
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
fonttools==4.57.0
idna==3.10
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
tflite-runtime==2.14.0
tifffile==2025.3.30
typing_extensions==4.13.2
urllib3==2.4.0
